requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.8.1
python-multipart==0.0.20
redis==5.2.1
email-validator==2.2.0
